A review app generator
This introduces a generator, `suspenders:staging:pull_requests`, for
setting up staging apps based for each pull request, specifically using
Heroku Review Apps.

This is namespaced under `staging` instead of `production` to reflect
that it specifically deals with staging apps. However, it is in practice
tied to the Heroku Pipeline.